+++ src/types/context.ts
@@ -93,6 +93,8 @@ export type FrameContext<
   _state = env['State'],
 > = Context<env, path, _state> & {
   /**
+   * @deprecated As of `v0.5.0`, this property is redundant (there is now only one render cycle) and will be removed in a future version.
+   *
    * Current render cycle of the frame.
    *
    * - `main` - Render cycle for the main frame route.
@@ -115,16 +117,6 @@ export type FrameContext<
    * - Ethereum: a transaction hash
    */
   transactionId?: FrameData['transactionId'] | undefined
-}
-
-export type FrameQueryContext<
-  env extends Env = Env,
-  path extends string = string,
-  //
-  _state = env['State'],
-> = Omit<FrameContext<env, path, _state>, 'req'> & {
-  req: undefined
-  state: _state
 }
 
 export type TransactionContext<